仲裁通模块对接接口

+++ library/src/main/java/com/hm/arbitrament/bean/GetArbCostResBean.java
@@ -16,8 +16,8 @@ public class GetArbCostResBean {
      * arbMoney : 0
      */
 
-    private int arbCost;
-    private int arbMoney;
+    private Double arbCost;
+    private Double arbMoney;
 
 
 }
